salesflow-app/backend/requirements.txt
# Sales Flow AI - Backend Dependencies
# Python 3.11+

# FastAPI Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Database
supabase>=2.3.0
sqlalchemy>=2.0.25
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
pyjwt>=2.8.0

# Cache
redis>=5.0.1

# AI / OpenAI
openai>=1.12.0
tiktoken>=0.5.2

# Validation & Settings
pydantic>=2.5.3
pydantic-settings>=2.1.0

# HTTP Client
httpx>=0.24.0
aiohttp>=3.9.0

# Utils
python-dotenv>=1.0.0
orjson>=3.9.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

